Dev yfl

See merge request elink/elink-juliet-flow!2

+++ elink-juliet-flow-domain/src/main/java/com/juliet/flow/repository/impl/FlowRepositoryImpl.java
@@ -312,7 +312,9 @@ public class FlowRepositoryImpl implements FlowRepository {
             .stream()
             .map(FlowEntity::getId)
             .collect(Collectors.toList());
-        nodeDao.delete(Wrappers.<NodeEntity>lambdaQuery().in(NodeEntity::getFlowId, idList));
+        if (CollectionUtils.isNotEmpty(idList)) {
+            nodeDao.delete(Wrappers.<NodeEntity>lambdaQuery().in(NodeEntity::getFlowId, idList));
+        }
     }
 
     private void deleteNodes(List<Node> nodes) {
